Change Project structure
In this project restrucutrarion it has been created a new class BaseAPIClient
with the common functions used later in each API module file like AthleteAPIClient
and ActivityAPIClient create a final class ApiManager that proceed all the different functions
in the modules Classes.

This has been done for a better organitzation of the Project.